feat: added reference to email exposure through the NPM API
The current [NPM API](https://registry.npmjs.org) is exposing under the `maintainers` and `_npmUser` properties the email addresses of the package maintainers and the author of the release publication. These extra fields were not part of the API in the past and most Npm users are not aware of this change.

Signed-off-by: Ulises Gascón <UlisesGascon@users.noreply.github.com>